requirements.txt
# AISTATEwww 

# Core engines 
openai-whisper>=20231117
pyannote.audio==4.0.2
huggingface_hub>=0.36.0,<1.0
torch>=2.2
torchaudio>=2.2
soundfile>=0.12.1
numpy>=1.26
tqdm>=4.66
reportlab>=4.0


# Web
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
python-multipart>=0.0.9

# --- Analysis tab (Ollama + documents) ---
httpx>=0.27.0
python-docx>=1.1.2
docxtpl>=0.18.0
openpyxl>=3.1.5
python-pptx>=0.6.23
PyPDF2>=3.0.1
pdfplumber>=0.11.4
PyMuPDF>=1.24.0
pytesseract>=0.3.10
Pillow>=10.4.0
pandas>=2.2.2
markdown>=3.6

# Translation module dependencies

# Core translation
transformers>=4.35.0
sentencepiece>=0.1.99
torch>=2.0.0
accelerate>=0.24.0

# Document processing
python-docx>=1.1.0
pdfplumber>=0.10.3
reportlab>=4.0.7

# Language detection
langdetect>=1.0.9

# Encryption
cryptography>=42.0
argon2-cffi>=23.1
